feat: adicionando a funçao formataCustoLogistico em um arquivo de utilidades
